requirements.txt
fastapi
uvicorn
supabase
pydantic
pydantic-settings
google-generativeai
groq
python-dotenv
